refactor: aleo config type
