+++ .changeset/stale-grapes-study.md
@@ -1,0 +1,8 @@
+---
+"@ledgerhq/coin-aleo": minor
+"ledger-live-desktop": minor
+"@ledgerhq/live-common": minor
+---
+
+refactor: aleo config type
+chore: more generic name for aleo learn more url
