+++ libs/coin-modules/coin-aleo/src/config.ts
@@ -1,18 +1,5 @@
-import buildCoinConfig, { type CurrencyConfig } from "@ledgerhq/coin-framework/config";
-import type { TransactionType } from "./types";
-
-export type AleoConfig = {
-  networkType: "mainnet" | "testnet";
-  apiUrls: {
-    node: string;
-    sdk: string;
-  };
-  feeByTransactionType: Record<TransactionType, number>;
-  feeSafetyMultiplier: number;
-  isFeeSponsored: boolean;
-};
-
-export type AleoCoinConfig = CurrencyConfig & AleoConfig;
+import buildCoinConfig from "@ledgerhq/coin-framework/config";
+import type { AleoCoinConfig } from "./types";
 
 const coinConfig = buildCoinConfig<AleoCoinConfig>();
