+++ libs/coin-modules/coin-aleo/src/index.ts
@@ -1,3 +1,1 @@
 export * from "./types";
-
-export type { AleoCoinConfig } from "./config";
