+++ libs/coin-modules/coin-aleo/src/types/index.ts
@@ -1,5 +1,6 @@
 export * from "./api";
 export * from "./bridge";
+export * from "./config";
 export * from "./logic";
 export * from "./sdk";
 export * from "./signer";
