+++ libs/ledger-live-common/src/families/aleo/setup.ts
@@ -5,11 +5,14 @@ import type { DeviceManagementKit } from "@ledgerhq/device-management-kit";
 import type { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 import { DmkSignerAleo } from "@ledgerhq/live-signer-aleo";
 import { createBridges } from "@ledgerhq/coin-aleo/bridge/index";
-import type { AleoCoinConfig } from "@ledgerhq/coin-aleo/config";
 import makeCliTools from "@ledgerhq/coin-aleo/test/cli";
 import aleoAddressResolver from "@ledgerhq/coin-aleo/signer/getAddress";
 import aleoViewKeyResolver from "@ledgerhq/coin-aleo/signer/getViewKey";
-import type { AleoSigner, Transaction as AleoTransaction } from "@ledgerhq/coin-aleo/types/index";
+import type {
+  AleoSigner,
+  Transaction as AleoTransaction,
+  AleoCoinConfig,
+} from "@ledgerhq/coin-aleo/types";
 import type Transport from "@ledgerhq/hw-transport";
 import type { Bridge } from "@ledgerhq/types-live";
 import { createResolver, executeWithSigner, type CreateSigner } from "../../bridge/setup";
